requirements.txt
google-genai
python-dotenv
pandas
pillow
requests
matplotlib
pytesseract